customize user

+++ src/Modules/vehicles/vehicle.controler.ts
@@ -1,18 +1,19 @@
 import { Request, Response } from "express";
 import { vehicleServes } from "./vehicle.serves";
+import { JwtPayload } from "jsonwebtoken";
 
 const getVehicles = async (req:Request,res:Response)=>{
     try{
         const result = await vehicleServes.getVehicle()
         res.status(200).json({
             success: true,
-            messatge: "Successfully Gotten All Vehicles",
+            message: "Vehicles retrieved successfully",
             details: result.rows
         })
     }catch(err:any){
         res.status(500).json({
             success: false,
-            message: err.any
+            message: err.message
         })
     }
 }
@@ -20,31 +21,51 @@ const getSingleVehicles = async (req:Request,res:Response)=>{
     const id = Number(req.params.id)
     try{
         const result = await vehicleServes.getSingleVehicles(id)
-        res.status(200).json({
+        if(result.rowCount === 0){
+            return res.status(200).json({
+                success:true,
+                message: "No vehicles found",
+                data:[]
+            })
+        }
+        return res.status(200).json({
             success: true,
-            messatge: "Successfully Gotten Vehicles",
-            details: result.rows
+            message: "Vehicle retrieved successfully",
+            data: result.rows
         })
     }catch(err:any){
         res.status(500).json({
             success: false,
-            message: err.any
+            message: err.message
         })
     }
 }
 
 const createVehicles = async (req:Request,res:Response)=>{
-    console.log(req.body);
     try{
-        const result = await vehicleServes.createVehicles(req.body)
-        res.status(201).json({
+        if (!req.user) {
+      return res.status(401).json({
+        success: false,
+        message: "Unauthorized"
+      });
+    }
+        const {role}=req.user as JwtPayload;
+        if(role === 'admin'){
+            const result = await vehicleServes.createVehicles(req.body)
+        return res.status(201).json({
             success: true,
-            message: result.rows,
+            message: "Vehicle created successfully",
+            data: result.rows,
+        })
+        }
+        return res.status(403).json({
+            success: false,
+            message: "only admin can create"
         })
     }catch(err:any){
         res.status(500).json({
             success: false,
-            message: err.any
+            message: err.message
         })
     }
 }
@@ -52,27 +73,57 @@ const createVehicles = async (req:Request,res:Response)=>{
 const updateVehicles = async (req:Request,res:Response)=>{
     const id = Number(req.params.id)
     try{
-        const result = await vehicleServes.updateVehicles(req.body,id)
+        if (!req.user) {
+      return res.status(401).json({
+        success: false,
+        message: "Unauthorized"
+      });
+    }
+        const {role} = req.user as JwtPayload;
+        if(role === 'admin'){
+            const result = await vehicleServes.updateVehicles(req.body,id)
+        return res.status(200).json({
+            success: true,
+            message: 'Successfully Updated',
+            details: result.rows,
+        })
+        }
+        return res.status(403).json({
+                success: false,
+                message: 'Only Admin Can Update'
+            })
     }catch(err:any){
         res.status(500).json({
             success: false,
-            message: err.any
+            message: err.message
         })
     }
 }
 const deleteVehicles = async (req:Request,res:Response)=>{
     const id = Number(req.params.id)
     try{
-        const result = await vehicleServes.deleteVehicles(id)
-        res.status(200).json({
+        if (!req.user) {
+      return res.status(401).json({
+        success: false,
+        message: "Unauthorized"
+      });
+    }
+        const {role} = req.user as JwtPayload;
+        if(role === 'admin'){
+            await vehicleServes.deleteVehicles(id)
+        return res.status(200).json({
             success: true,
-            message: "Successfully Deleted",
-            detatils: result.rows,
+            message: "Vehicle deleted successfully"
+        })
+        }
+        return res.status(403).json({
+            success: false,
+            message: "only admin can delete"
         })
     }catch(err:any){
         res.status(500).json({
             success: false,
-            message: err.any
+            message: err.message
         })
     }
 }
